[level 0] Title: 정수 부분, Time: 0.02 ms, Memory: 74.7 MB -BaekjoonHub

+++ 프로그래머스/0/181850. 정수 부분/정수 부분.java
@@ -1,5 +1,7 @@
 class Solution {
     public int solution(double flo) {
-        return ((Double)flo).intValue();
+        // return ((Double)flo).intValue();
+        // return (int)Math.floor(flo);
+        return (int)flo;
     }
 }
